Add postprocessing option to update \pgfimage calls with correct directory.
The PGF backend has no knowledge of the project structure and has assume
the output figure is in the same directory as the main TeX file. As
such, all calls to \pgfimage are missing the directory information. We
can do a bit of postprocessing to add the directory into the filename.