Fix setup_figure() for specific dimensions; add unit tests.
This mainly involved splitting out the dimension parsing code in the
config parser so it could be called with a string rather than on a
section and option. The resulting logic in the setup_figure() function
is now a bit cleaner as well.

The unit tests call setup_figure() with all combinations of fractional
and specific dimensions.