Fix preamble handling; add tests for using custom fonts with preamble.
The current matplotlib.rcParams validator for the preamble splits the
string on commas. The resulting list is subsequently re-joined with
newlines, causing TeX to fail when the preamble contained commas (e.g.,
key-value configuration of a package). Two pull requests switching to a
proper TeX validator have been accepted and should be included when
matplotlib 3.1.0 is released:

    * https://github.com/matplotlib/matplotlib/pull/12674
    * https://github.com/matplotlib/matplotlib/pull/12805

In the meantime, we can work around this by replacing the validator with
a call to str.splitlines(). The workaround is implemented in a
future-proof manner as it is only installed if the TeX validator could
not be imported, i.e., as soon as it is used with a version of
matplotlib containing the validator the workaround should be ignored.

This also adds a couple of unit tests which use the preamble to set
custom fonts. The font used is Cotham Sans, released under the SIL Open
Font License 1.1.