Tracking: fix bugs mis-detecting written files as read files.
At some point the reported mode on the file objects has changed;
sometimes files opened with the 'w+b' argument to open() report a mode
of 'r+b'. This might be equivalent from the viewpoint of the code
writing the file, but broke the file tracking.

Now, we only attempt to track file objects that are subclasses of
io.IOBase (which most files should be) and use their writeable() and
readable() methods to determine their state.