Drop old preamble workaround.
The pull request making the workaround obsolete was merged 3 years ago.
As of Matplotlib 3.5.0 the function the workaround checked for no longer
exists so it failed again. Without the workaround 3.1+ is supported (and
possibly the last minor releases of the 3.0 series) so lets assume any
user has upgraded sometime in the last 3 years.